# gen/http/openapi.yaml
swagger: "2.0"
info:
  title: webcup
  description: Best API REST building with GoaDesign
  contact:
    name: Guillaume MORIN
    email: guillaume.morin@epitech.eu
    url: https://guillaume-morin.fr/
  license:
    name: GOA X GM
    url: https://guillaume-morin.fr/
  version: "1.0"
host: :0
consumes:
- application/json
- application/xml
- application/gob
produces:
- application/json
- application/xml
- application/gob
paths:
...
  /public/{path}:
    get:
      tags:
      - fileapi
      summary: Download bin/public
      description: Serve static content.
      operationId: fileapi#/public/{*path}
      parameters:
      - name: path
        in: path
        description: Relative file path
        required: true
        type: string
      responses:
        "200":
          description: File downloaded
          schema:
            type: file
        "404":
          description: File not found
          schema:
            $ref: '#/definitions/Error'
      schemes:
      - http
...
  /v1/bo/file/add:
    post:
      tags:
      - files
      summary: importFile files
      description: Import file
      operationId: files#importFile
      consumes:
      - multipart/form-data
      parameters:
      - name: Authorization
        in: header
        description: OAuth token
        required: false
        type: string
        pattern: ^Bearer [^ ]+$
      - name: jwtToken
        in: header
        description: Jwt token
        required: false
        type: string
        pattern: ^Bearer [^ ]+$
      - name: ImportFileRequestBody
        in: formData
        required: true
        schema:
          $ref: '#/definitions/FilesImportFileRequestBody'
          required:
          - filename
          - content
          - format
      responses:
        "201":
          description: Created response.
          schema:
            $ref: '#/definitions/FilesImportFileResponseBody'
            required:
            - success
            - file
        "500":
          description: Internal Server Error response.
          schema:
            $ref: '#/definitions/FilesImportFileUnknownErrorResponseBody'
            required:
            - err
            - success
            - error_code
      schemes:
      - http
      security:
      - OAuth2_header_Authorization: []
        jwt_header_jwtToken: []
  /v1/bo/file/remove:
    patch:
      tags:
      - files
      summary: deleteFile files
      description: Delete one file by ID
      operationId: files#deleteFile
      parameters:
      - name: Authorization
        in: header
        description: OAuth token
        required: false
        type: string
        pattern: ^Bearer [^ ]+$
      - name: jwtToken
        in: header
        description: Jwt token
        required: false
        type: string
        pattern: ^Bearer [^ ]+$
      - name: DeleteFileRequestBody
        in: body
        required: true
        schema:
          $ref: '#/definitions/FilesDeleteFileRequestBody'
          required:
          - url
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/FilesDeleteFileResponseBody'
            required:
            - success
        "500":
          description: Internal Server Error response.
          schema:
            $ref: '#/definitions/FilesDeleteFileUnknownErrorResponseBody'
            required:
            - err
            - success
            - error_code
      schemes:
      - http
      security:
      - OAuth2_header_Authorization: []
        jwt_header_jwtToken: []
...
definitions:
...
  Error:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: true
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the
          problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence
          of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: true
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: false
    description: Error response result type (default view)
    example:
      fault: false
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: true
      timeout: true
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
  FilesDeleteFileRequestBody:
    title: FilesDeleteFileRequestBody
    type: object
    properties:
      url:
        type: string
        example: /public/uploads/2021/12/2ca51d10-b660-4b2c-b27f-f7a119642885.png
    example:
      url: /public/uploads/2021/12/2ca51d10-b660-4b2c-b27f-f7a119642885.png
    required:
    - url
  FilesDeleteFileResponseBody:
    title: FilesDeleteFileResponseBody
    type: object
    properties:
      success:
        type: boolean
        example: true
    example:
      success: true
    required:
    - success
  FilesDeleteFileUnknownErrorResponseBody:
    title: FilesDeleteFileUnknownErrorResponseBody
    type: object
    properties:
      err:
        type: string
        example: sql no rows affected
      error_code:
        type: string
        example: TX_UPDATE_ITEM
      success:
        type: boolean
        default: false
        example: false
    description: Error not identified (500)
    example:
      err: sql no rows affected
      error_code: TX_UPDATE_ITEM
      success: true
    required:
    - err
    - success
    - error_code
  FilesImportFileRequestBody:
    title: FilesImportFileRequestBody
    type: object
    properties:
      content:
        type: string
        description: content of image
        example:
        - 72
        - 105
        - 99
        - 32
        - 100
        - 101
        - 98
        - 105
        - 116
        - 105
        - 115
        - 32
        - 97
        - 117
        - 116
        - 32
        - 105
        - 112
        - 115
        - 117
        - 109
        - 32
        - 101
        - 116
        - 46
        format: byte
      filename:
        type: string
        description: uploaded file name
        example: foo.jpg
      format:
        type: string
        description: uploaded file format
        example: image/jpeg
      h:
        type: integer
        description: height of image if you crop
        example: 8060800046913052297
        format: int64
      mime:
        type: string
        description: url file
        example: Et provident iusto quo a sit.
      size:
        type: integer
        description: size of image
        example: 735968203255190022
        format: int64
      url:
        type: string
        description: url file
        example: Animi consequatur enim.
      w:
        type: integer
        description: width of image if you crop
        example: 5402964006909297550
        format: int64
    example:
      content:
      - 68
      - 105
      - 103
      - 110
      - 105
      - 115
      - 115
      - 105
      - 109
      - 111
      - 115
      - 32
      - 101
      - 97
      - 46
      filename: foo.jpg
      format: image/jpeg
      h: 2853712597827717815
      mime: Corrupti sed voluptatum enim.
      size: 8637745267229013526
      url: In voluptas.
      w: 3984455343309882416
    required:
    - filename
    - content
    - format
  FilesImportFileResponseBody:
    title: FilesImportFileResponseBody
    type: object
    properties:
      file:
        $ref: '#/definitions/resFileResponseBody'
      success:
        type: boolean
        example: false
    example:
      file:
        id: 5dfb0bf7-597a-4250-b7ad-63a43ff59c25
        mime: Tenetur aut est voluptatibus fuga.
        name: foo.png
        size: 6477760254985632700
        url: Alias sit iure ut.
      success: false
    required:
    - success
    - file
  FilesImportFileUnknownErrorResponseBody:
    title: FilesImportFileUnknownErrorResponseBody
    type: object
    properties:
      err:
        type: string
        example: sql no rows affected
      error_code:
        type: string
        example: TX_UPDATE_ITEM
      success:
        type: boolean
        default: false
        example: false
    description: Error not identified (500)
    example:
      err: sql no rows affected
      error_code: TX_UPDATE_ITEM
      success: true
    required:
    - err
    - success
    - error_code
...
  resFileResponseBody:
    title: resFileResponseBody
    type: object
    properties:
      id:
        type: string
        example: 5dfb0bf7-597a-4250-b7ad-63a43ff59c25
        format: uuid
      mime:
        type: string
        example: Beatae molestiae sed rerum non natus libero.
      name:
        type: string
        example: foo.png
      size:
        type: integer
        example: 3621460537578703638
        format: int64
      url:
        type: string
        example: Perferendis laboriosam quae unde omnis.
    example:
      id: 5dfb0bf7-597a-4250-b7ad-63a43ff59c25
      mime: Deleniti dolores voluptatibus iusto fugiat sed.
      name: foo.png
      size: 986685068245262051
      url: Earum quia id nisi totam et sint.
    required:
    - id
    - name
    - url
...
securityDefinitions:
  OAuth2_header_Authorization:
    type: oauth2
    description: Use OAuth2 to authenticate
    flow: application
    tokenUrl: /authorization
    scopes:
      api:read: Provides read access
  jwt_header_jwtToken:
    type: apiKey
    description: |-
      Secures endpoint by requiring a valid JWT token retrieved via the signin endpoint. Supports scopes "api:read" and "api:write".

      **Security Scopes**:
        * `api:read`: Read-only access
        * `api:write`: Read and write access
    name: jwtToken
    in: header